# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "DeviceRequest": {
        "additionalProperties": false,
        "properties": {
          "app_version": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version"
          },
          "client_device_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Device Id"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "platform": {
            "anyOf": [
              {
                "enum": [
                  "ios",
                  "android",
                  "web",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          }
        },
        "title": "DeviceRequest",
        "type": "object"
      },
      "EmailChallengeRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 320,
            "pattern": "^[^@\\s]{1,64}@[^@\\s]+\\.[^@\\s]{2,}$",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "EmailChallengeRequest",
        "type": "object"
      },
      "EmailChallengeResponse": {
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "title": "Challenge Id",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "expires_at"
        ],
        "title": "EmailChallengeResponse",
        "type": "object"
      },
      "EmailVerifyRequest": {
        "additionalProperties": false,
        "description": "Either ``challenge_id`` + ``code`` from the email, or the magic-link ``link_token``.",
        "properties": {
          "challenge_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Challenge Id"
          },
          "code": {
            "anyOf": [
              {
                "pattern": "^[0-9]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "link_token": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Token"
          },
          "merge_guest_participations": {
            "default": false,
            "title": "Merge Guest Participations",
            "type": "boolean"
          }
        },
        "title": "EmailVerifyRequest",
        "type": "object"
      },
      "ExternalSignInRequest": {
        "additionalProperties": false,
        "description": "Google/Apple ID token. With a guest bearer token this upgrades or claims the guest;\nwith a registered bearer token it re-authenticates (step-up) the current session.",
        "properties": {
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "id_token": {
            "maxLength": 8192,
            "minLength": 20,
            "title": "Id Token",
            "type": "string"
          },
          "merge_guest_participations": {
            "default": false,
            "title": "Merge Guest Participations",
            "type": "boolean"
          },
          "nonce": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nonce"
          }
        },
        "required": [
          "id_token"
        ],
        "title": "ExternalSignInRequest",
        "type": "object"
      },
...
      "JwksResponse": {
        "properties": {
          "keys": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Keys",
            "type": "array"
          }
        },
        "required": [
          "keys"
        ],
        "title": "JwksResponse",
        "type": "object"
      },
...
      "PublicProblemResponse": {
        "description": "Typed-client-safe problem fields guaranteed across every API error.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "PublicProblemResponse",
        "type": "object"
      },
...
      "RefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "maxLength": 256,
            "minLength": 20,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present when a new session starts or rotates; absent after step-up.",
            "title": "Refresh Token"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "token_type": {
            "const": "Bearer",
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfileResponse"
          }
        },
        "required": [
          "access_token",
          "expires_at",
          "refresh_token",
          "session_id",
          "user"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserProfileResponse": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "registered",
              "guest"
            ],
            "title": "Kind",
            "type": "string"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "display_name",
          "email",
          "locale",
          "timezone",
          "version"
        ],
        "title": "UserProfileResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for plans with your people.",
    "title": "Beluno API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/.well-known/jwks.json": {
      "get": {
        "operationId": "signing_keys__well_known_jwks_json_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JwksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "summary": "Signing Keys",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/auth/apple": {
      "post": {
        "operationId": "sign_in_with_apple_v1_auth_apple_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Too many requests; retry after the Retry-After interval."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sign In With Apple",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/email/challenges": {
      "post": {
        "description": "Always accepted for a well-formed address; never reveals whether an account exists.",
        "operationId": "start_email_challenge_v1_auth_email_challenges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailChallengeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailChallengeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Too many requests; retry after the Retry-After interval."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "summary": "Start Email Challenge",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/email/verify": {
      "post": {
        "operationId": "verify_email_challenge_v1_auth_email_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Too many requests; retry after the Retry-After interval."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Verify Email Challenge",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/google": {
      "post": {
        "operationId": "sign_in_with_google_v1_auth_google_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Too many requests; retry after the Retry-After interval."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sign In With Google",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_tokens_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Too many requests; retry after the Retry-After interval."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "summary": "Refresh Tokens",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}